Module vectis_govec::contract

Enums

Role

Functions

create_accounts

execute

execute_burn

Burning of Govec must be done by dao

execute_exit

Exits the DAO remove the requester from the ledger any balance in the requester account will be sent to the dao

execute_mint

execute_send

Send can be used for the staking contract and other contracts in the future For delegation, use transfer instead where only wallets are in the whitelist

execute_transfer

execute_transfer_deposit

This is exclusive used by proposal module

execute_update_config_addr

execute_update_mint_amount

execute_update_mint_cap

execute_update_staking

govec_execute_update_marketing

govec_execute_upload_logo

instantiate

query

query_balance

query_balance_joined

query_contract_addr

query_dao

query_dao_tunnel

query_factory

query_mint_amount

query_minter

query_staking

query_token_info